component renaming
