+++ addon/components/eui-notification-badge/index.hbs
@@ -1,5 +1,5 @@
 <span class={{class-names
-  componentName="EuiBadgeNotification"
+  componentName="EuiNotificationBadge"
   size=(arg-or-default @size "s")
   color=(arg-or-default @color "accent")}}
   ...attributes
