+++ addon/utils/css-mappings/eui-notification-badge.ts
