--- app/components/eui-badge-notification/index.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-eui/components/eui-badge-notification';
