+++ app/components/eui-notification-badge/index.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-eui/components/eui-notification-badge';
