+++ tests/dummy/app/templates/demo/badge.hbs
@@ -84,11 +84,11 @@
     <EuiTitle @tag="h2" @size="m">
       Notification Badge
     </EuiTitle>
-    <EuiBadgeNotification>1</EuiBadgeNotification>
-    <EuiBadgeNotification>2</EuiBadgeNotification>
-    <EuiBadgeNotification>3</EuiBadgeNotification>
-    <EuiBadgeNotification @color="subdued">4</EuiBadgeNotification>
-    <EuiBadgeNotification @size="m">5</EuiBadgeNotification>
+    <EuiNotificationBadge>1</EuiNotificationBadge>
+    <EuiNotificationBadge>2</EuiNotificationBadge>
+    <EuiNotificationBadge>3</EuiNotificationBadge>
+    <EuiNotificationBadge @color="subdued">4</EuiNotificationBadge>
+    <EuiNotificationBadge @size="m">5</EuiNotificationBadge>
     <EuiSpacer />
 
     <EuiTitle @size="m">
